FPVTL-2429: POC for Bundles using Category ID
