+++ src/test/java/uk/gov/hmcts/reform/prl/mapper/bundle/BundleCreateRequestMapperTest.java
@@ -1,9 +1,9 @@
 package uk.gov.hmcts.reform.prl.mapper.bundle;
 
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.prl.constants.PrlAppsConstants;
 import uk.gov.hmcts.reform.prl.enums.DocTypeOtherDocumentsEnum;
@@ -94,9 +94,16 @@ import static uk.gov.hmcts.reform.prl.utils.ElementUtils.wrapElements;
 
 @RunWith(MockitoJUnitRunner.class)
 public class BundleCreateRequestMapperTest {
-    @InjectMocks
+
+    private HearingDetailsMapperUtil hearingDetailsMapperUtil;
+
     private BundleCreateRequestMapper bundleCreateRequestMapper;
 
+    @Before
+    public void setUp() throws Exception {
+        hearingDetailsMapperUtil = new HearingDetailsMapperUtil();
+        bundleCreateRequestMapper = new BundleCreateRequestMapper(hearingDetailsMapperUtil);
+    }
 
     @Test
     public void testAddOtherDocument_excludesRedactedByUrl() {
